Move '$ rio edit-info' tests into their own file.
